Restructure the ClientInfo-Move-Respawn problem
So this "fixes" the conflict of how ClientInfo, Move, and subsequently
Respawn should interact with one another. It also includes a trivial
change to the net module that is simply a convenient new syntax for
pushing packets.

The Respawn problem turns out to be a non-issue, the normal interaction
of ClientInfo and Move turns out to be sufficient for Respawn. Maybe
not 100% the correct behavior, but it works fine so I'm going to let it
go. Note: Should check out if it's ok to send Position Updates while
dead

ClientInfo and move... IMHO the way this should work is Physics and Move
work together to come up with a movement plan, they then update
ClientInfo with new positions (not sure if Move should be solely
responsible for this). Every client_tick, Move is responsible for
sending the position currently held by ClientInfo to the server.