Got rid of Client class, moved settings to plugin
This was a long time coming, settings parsing and handling arguably doesn't even
belong in Spock (we're "just a framework"). However, since the typical use case
involves running Spock basically out of the demo script, it makes sense that we
provide some sort of settings parsing. This commit lays the foundation by
getting settings handling out of the Client class stub (finally removing the
stub once and for all) and moving it to a specially handled settings plugin.

Settings plugins are a special case and don't follow the usual pl_announce
semantics for announcing and providing functionality. Instead they need to be
passed at initialization time as a kwarg, 'settings_mixin'. Otherwise the
PluginLoader defaults to the plugins/core SettingsPlugin.