Implement new event PACKET_ERR
PACKET_ERR is a new event emitted by net that allows plugins to hook
into failed packet decodes, as well as peak inside the packet buffer.
Nothing is guaranteed about the packet or the packet buffer, but this
can be very useful for debugging to inspect a packet that has decoded
mostly correctly.