A great of of work implementing notchian physics
This commit is the result of a lot of work from myself and gamingrobot
to reverse engineer notchian client physics. As such, it's a massive
rework of the existing physics plugin. A lot of behavior is still
missing, and there is still some aberrant behavior. But this commit is
a great deal of progress towards correct physics.